src/qe_sac/gnn_encoder.py
"""
GNN Encoder — topology-aware replacement for the flat MLP CAE.

Architecture:
    Node features : [V_pu, P_load_norm, Q_load_norm]  (3 per bus)
    Edge features : [r, x]  (resistance, reactance per line)
    Message passing: 2 rounds of GCNConv
    Global pool → 8-dim latent (tanh * π to match VQC input range)

Why GNN over flat MLP CAE:
    - Knows bus topology (bus 3 connects to bus 4)
    - Same model works on 13-bus, 34-bus, 123-bus without retraining
    - Physically meaningful: voltage information propagates through
      graph edges — exactly how power flows on the grid

IEEE 13-bus graph (edges defined by line connections):
    sourcebus - 650 - 632 - 670 - 671 - 680
                        \\         \\         \\
                        633-634   684-611    692-675
                                    \\
                                    652

Reference:
    Kipf & Welling (2017) "Semi-supervised Classification with GCN"
    Lin et al. (2025) (base paper, MLP CAE) — this replaces their encoder
"""

from __future__ import annotations

import numpy as np
import torch
import torch.nn as nn
import torch.nn.functional as F

try:
    from torch_geometric.nn import GCNConv, global_mean_pool
    from torch_geometric.data import Data, Batch
    _HAS_PYG = True
except ImportError:
    _HAS_PYG = False


# ── IEEE 13-bus edge list (undirected, 0-indexed bus order) ──────────────────
# Bus order: [sourcebus, 650, 632, 670, 671, 680, 633, 634, 684, 611, 692, 675, 652, 645, 646]
# Index:      [     0,     1,   2,   3,   4,   5,   6,   7,   8,   9,  10,  11,  12,  13,  14]

_EDGE_INDEX_13BUS = torch.tensor([
    # from, to
    [0, 1], [1, 0],   # sourcebus - 650
    [1, 2], [2, 1],   # 650 - 632
    [2, 3], [3, 2],   # 632 - 670
    [3, 4], [4, 3],   # 670 - 671
    [4, 5], [5, 4],   # 671 - 680
    [2, 6], [6, 2],   # 632 - 633
    [6, 7], [7, 6],   # 633 - 634
    [4, 8], [8, 4],   # 671 - 684
    [8, 9], [9, 8],   # 684 - 611
    [4, 10],[10, 4],  # 671 - 692
    [10,11],[11,10],  # 692 - 675
    [8, 12],[12, 8],  # 684 - 652
    [2, 13],[13, 2],  # 632 - 645
    [13,14],[14,13],  # 645 - 646
], dtype=torch.long).t().contiguous()  # shape (2, 28)

# Line impedance (r, x) per edge — approximate values from IEEE 13-bus data
# Order matches _EDGE_INDEX_13BUS (undirected, duplicate for both directions)
_EDGE_ATTR_13BUS = torch.tensor([
    [0.000, 0.000], [0.000, 0.000],  # sourcebus-650 (transformer)
    [0.347, 1.018], [0.347, 1.018],  # 650-632 (mtx601, 2000ft)
    [0.347, 1.018], [0.347, 1.018],  # 632-670 (mtx601, 667ft)
    [0.347, 1.018], [0.347, 1.018],  # 670-671 (mtx601, 1333ft)
    [0.347, 1.018], [0.347, 1.018],  # 671-680 (mtx601, 1000ft)
    [0.753, 1.181], [0.753, 1.181],  # 632-633 (mtx602, 500ft)
    [0.753, 1.181], [0.753, 1.181],  # 633-634 (mtx602, 500ft)
    [0.347, 1.018], [0.347, 1.018],  # 671-684 (mtx601, 300ft)
    [0.753, 1.181], [0.753, 1.181],  # 684-611 (mtx602, 300ft)
    [0.347, 1.018], [0.347, 1.018],  # 671-692 (mtx601, 1ft)
    [0.347, 1.018], [0.347, 1.018],  # 692-675 (mtx601, 500ft)
    [0.753, 1.181], [0.753, 1.181],  # 684-652 (mtx602, 800ft)
    [0.347, 1.018], [0.347, 1.018],  # 632-645 (mtx602)
    [0.347, 1.018], [0.347, 1.018],  # 645-646 (mtx602)
], dtype=torch.float32)  # shape (28, 2)

N_BUSES_13 = 15


class GNNEncoder(nn.Module):
    """
    GNN-based state encoder for distribution grid VVC.

    Replaces the flat MLP CAE from Lin et al. (2025).
    Takes per-bus features [V, P, Q] and edge features [r, x].
    Outputs 8-dim latent in [-π, π] — same interface as CAE.encode().

    Parameters
    ----------
    node_feat_dim  : features per bus (default 3: V, P, Q)
    edge_feat_dim  : features per line (default 2: r, x)
    hidden_dim     : GCN hidden dimension
    latent_dim     : output latent dimension (must be 8 for VQC)
    n_buses        : number of buses in the feeder
    edge_index     : (2, E) tensor of edge connectivity
    edge_attr      : (E, edge_feat_dim) tensor of line parameters
    """

    def __init__(
        self,
        node_feat_dim: int = 3,
        edge_feat_dim: int = 2,
        hidden_dim:    int = 32,
        latent_dim:    int = 8,
        n_buses:       int = N_BUSES_13,
        edge_index:    torch.Tensor | None = None,
        edge_attr:     torch.Tensor | None = None,
    ):
        if not _HAS_PYG:
            raise ImportError(
                "torch-geometric is required for GNNEncoder. "
                "Install with: pip install torch-geometric"
            )
        super().__init__()
        self.n_buses   = n_buses
        self.latent_dim = latent_dim

        # Register graph structure as buffers (not trained, but saved in state_dict)
        ei = edge_index if edge_index is not None else _EDGE_INDEX_13BUS
        ea = edge_attr  if edge_attr  is not None else _EDGE_ATTR_13BUS
        self.register_buffer("edge_index", ei)
        self.register_buffer("edge_attr",  ea)

        # Edge feature → node feature projection (fused into node update)
        # We project [node_feat + edge_feat] instead of standard GCN
        self.edge_proj = nn.Linear(edge_feat_dim, node_feat_dim)

        # GCN layers
        self.conv1 = GCNConv(node_feat_dim, hidden_dim)
        self.conv2 = GCNConv(hidden_dim, hidden_dim)

        # Global pool → latent
        self.fc_latent = nn.Linear(hidden_dim, latent_dim)

    def _obs_to_graph(self, obs: torch.Tensor) -> "Batch":
        """
        Convert flat observation vector to PyG Batch.

        For 13-bus (42-dim DistFlow):
            obs[0:13]  = V magnitudes
            obs[13:26] = P loads (normalised)
            obs[26:39] = Q loads (normalised)
            obs[39:41] = cap status
            obs[41]    = tap position

        Node features = [V, P, Q] per bus.
        Global features (caps, tap) are appended to graph-level features.
        """
        if obs.dim() == 1:
            obs = obs.unsqueeze(0)

        batch_size = obs.shape[0]
        n = self.n_buses
        graphs = []

        for b in range(batch_size):
            o = obs[b]
            # Extract per-bus features — handle both 42-dim and 93-dim obs
            if o.shape[0] >= 3 * n:
                v = o[:n]
                p = o[n:2*n]
                q = o[2*n:3*n]
            else:
                # Fallback: replicate v for p and q
                v = o[:n]
                p = torch.zeros(n, device=obs.device)
                q = torch.zeros(n, device=obs.device)

            node_x = torch.stack([v, p, q], dim=1)  # (n_buses, 3)
            graphs.append(Data(
                x=node_x,
                edge_index=self.edge_index,
                edge_attr=self.edge_attr,
            ))

        return Batch.from_data_list(graphs)

    def encode(self, obs: torch.Tensor) -> torch.Tensor:
        """
        Encode observation to 8-dim latent in [-π, π].

        Parameters
        ----------
        obs : (obs_dim,) or (batch, obs_dim)

        Returns
        -------
        latent : (8,) or (batch, 8)  in [-π, π]
        """
        single = obs.dim() == 1
        batch = self._obs_to_graph(obs)

        # GCN message passing
        x = F.relu(self.conv1(batch.x, batch.edge_index))
        x = F.relu(self.conv2(x, batch.edge_index))

        # Global mean pool → per-graph representation
        x = global_mean_pool(x, batch.batch)  # (batch_size, hidden_dim)

        # Project to latent
        latent = torch.tanh(self.fc_latent(x)) * torch.pi  # (batch_size, latent_dim)

        return latent.squeeze(0) if single else latent

    def forward(self, obs: torch.Tensor) -> tuple[torch.Tensor, torch.Tensor]:
        """
        Forward pass matching CAE interface: returns (reconstruction, latent).
        GNN encoder has no reconstruction — returns (zeros, latent).
        """
        latent = self.encode(obs)
        # No reconstruction in GNN mode — return zeros placeholder
        if obs.dim() == 1:
            recon = torch.zeros_like(obs)
        else:
            recon = torch.zeros_like(obs)
        return recon, latent


def train_gnn_encoder(
    encoder: GNNEncoder,
    obs_data: np.ndarray,
    n_steps: int = 100,
    lr: float = 1e-3,
    device: str = "cpu",
) -> float:
    """
    Pretrain GNN encoder with a reconstruction proxy loss.

    Since GNN has no decoder, we use a contrastive smoothness loss:
        neighbouring buses (connected by edges) should have similar latent.
    This encourages the GNN to produce physically meaningful embeddings.

    Returns final loss value.
    """
    encoder = encoder.to(device)
    optimizer = torch.optim.Adam(encoder.parameters(), lr=lr)
    data = torch.tensor(obs_data, dtype=torch.float32, device=device)

    final_loss = 0.0
    for step in range(n_steps):
        idx = np.random.randint(0, len(data), min(64, len(data)))
        batch_obs = data[idx]

        latent = encoder.encode(batch_obs)  # (batch, 8)

        # Smoothness loss: latent should not be all the same (avoid collapse)
        # Use variance regularisation to prevent mode collapse
        var_loss = -latent.var(dim=0).mean()   # maximise variance across batch

        # L2 regularisation to keep latent bounded
        reg_loss = (latent ** 2).mean() * 0.01

        loss = var_loss + reg_loss
        optimizer.zero_grad()
        loss.backward()
        optimizer.step()
        final_loss = float(loss)

    return final_loss
