requirements.txt
numpy==1.26.4
pandas==2.2.3
scipy==1.13.1
pandapower==2.14.11
numba==0.61.2
llvmlite==0.44.0
matplotlib==3.10.8
torch==2.5.1
stable-baselines3==2.5.0
gymnasium==1.0.0
jupyter==1.1.1
ipykernel==6.29.5
nbformat==5.10.4
# QE-SAC dependencies
pennylane
pennylane-lightning
powergym
opendssdirect.py
